view: Replace multi* libraries with base-x.
This pertains to the logic needed to extract the prefixless-sha1-
multihash from a cid for an IPLD git commit object, in the Filetree
view component.

Using multibase, multicodec, and multihashes is problematic because
they are not compiled to ES5, and I haven't been able to figure out
how to achieve this with webpack yet.

Fortunately, multicodec and multihashes are unnecessary, because for
this application, the codec should always be 'git-raw', and the
multihash will always be sha1-based. However, re: multibase, this
commit does not add logic for checking the multibase table for all
possible bases one can use for a CID, and instead introduces a hard-
coded limitation of base58btc.